my-course/requirements.txt.txt
streamlit==1.26.0
pandas==2.0.3
python-docx==0.8.11
Pillow==10.0.0
